scripts/requirements.txt
rembg[new]>=2.0.0
Pillow>=10.0.0


